Rearrange modules a bit.
