+++ eliottree/__init__.py
@@ -1,8 +1,9 @@
 from eliottree._parse import tasks_from_iterable
+from eliottree._render import render_tasks
 from eliottree.filter import (
     filter_by_end_date, filter_by_jmespath, filter_by_start_date,
     filter_by_uuid)
-from eliottree.render import render_task_nodes, render_tasks
+from eliottree.render import render_task_nodes
 from eliottree.tree import Tree
